shift value can fit in 8 bits

+++ engine.c
@@ -148,7 +148,7 @@ static int32_t scale(int32_t a, int32_t b, int32_t c)
 }
 
 
-static int32_t mulscale(int32_t a, int32_t b, int32_t c)
+static int32_t mulscale(int32_t a, int32_t b, uint_fast8_t c)
 {
 	return ((int64_t)a * b) >> c;
 }
@@ -179,7 +179,7 @@ static int32_t mulscale(int32_t a, int32_t b, int32_t c)
 #define mulscale32(a,b) mulscale((a),(b),32)
 
 
-static int32_t dmulscale(int32_t a, int32_t b, int32_t c, int32_t d, int32_t e)
+static int32_t dmulscale(int32_t a, int32_t b, int32_t c, int32_t d, uint_fast8_t e)
 {
 	return (((int64_t)a * b) + ((int64_t)c * d)) >> e;
 }
@@ -199,7 +199,7 @@ static int32_t dmulscale(int32_t a, int32_t b, int32_t c, int32_t d, int32_t e)
 #define dmulscale32(a,b,c,d) dmulscale((a),(b),(c),(d),32)
 
 
-static int32_t divscale(int32_t a, int32_t b, int32_t c)
+static int32_t divscale(int32_t a, int32_t b, uint_fast8_t c)
 {
 	return ((int64_t)a << c) / b;
 }
